Move all image-related functions into single file
