+++ components/CustomDesignsGrid.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import Link from "next/link";
 
 import { patterns } from "@/lib/config";
-import { getPatternThumbnailURL } from "@/lib/utils/cloudinary";
+import { getPatternThumbnailURL } from "@/lib/utils/image";
 
 import type { Pattern } from "@/lib/types/miscellaneous";
 
@@ -10,17 +10,18 @@ type Props = {
   category: Pattern["category"];
 };
 
-const MINIMUM_EMPTY_SLOTS = 28;
+const minimumEmptySlots = 28;
 
 export default function PatternsGrid({ category }: Props) {
   // Filtering patterns array to include only objects that have a same category to the category prop
   const filteredPatterns: Pattern[] = patterns.filter((pattern) => pattern.category === category);
 
+  // Calculate the number of placeholder slots to be filled
   const patternCount = filteredPatterns.length;
-  const emptySlots =
-    patternCount > MINIMUM_EMPTY_SLOTS
-      ? MINIMUM_EMPTY_SLOTS * Math.ceil(patternCount / MINIMUM_EMPTY_SLOTS) - patternCount
-      : MINIMUM_EMPTY_SLOTS - patternCount;
+  const placeholderCount =
+    patternCount > minimumEmptySlots
+      ? minimumEmptySlots * Math.ceil(patternCount / minimumEmptySlots) - patternCount
+      : minimumEmptySlots - patternCount;
 
   return (
     <div className="grid grid-cols-4 gap-4 md:grid-cols-7">
@@ -44,7 +45,7 @@ export default function PatternsGrid({ category }: Props) {
           />
         </Link>
       ))}
-      {[...Array(emptySlots)].map((_, i) => (
+      {[...Array<undefined>(placeholderCount)].map((_, i) => (
         <div className="flex aspect-square h-full w-full items-center justify-center" key={i}>
           <div className="aspect-square h-5.5 w-5.5 rounded-full bg-bone"></div>
         </div>
